consider local doesn't belong in exception notification - pulling out to optional module
