+++ init.rb
@@ -2,3 +2,4 @@ require "action_mailer"
 require "exception_notifier"
 require "exception_notifiable"
 require "exception_notifier_helper"
+require "consider_local"
